CHROMIUM: X86: Explicitly add -fomit-frame-pointer to boot/compressed.
We are changing the compiler default in ChromiumOS to fno-omit-frame-pointer.
The makefile under boot/compressed does not specify fomit-frame-pointer
directly so it gets the compiler default, which at the moment is
fomit-frame-pointer. Since we are going to change the default, we will
add the option explicitly to the Makefile. Not doing that causes the
kernel image to exceed its size limits.

BUG=chromium:464650
TEST=cbuildbot --hwtest amd64-generic-full
Signed-off-by: Luis A Lozano <llozano@chromium.org>

Change-Id: I77972fb23189c35fd42ae9002e9d203d23336ec4
Reviewed-on: https://chromium-review.googlesource.com/260433
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Conflicts:
	arch/x86/boot/compressed/Makefile
[rebase49(groeck): Context conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>