WIP: cpufreq/intel_pstate: Implement energy-efficient controller statistics and status calculation.
The goal of the helper code introduced here is to compute two
informational data structures: struct lp_sched_info aggregating
various scheduling statistics gathered in every call of the
update_util() hook, and struct lp_status_sample which contains status
information derived from the former indicating whether the system is
likely to have an IO or a CPU bottleneck.  This will be used as main
heuristic input by a new controller that assists the HWP at finding a
reasonably energy-efficient P-state given the additional information
available to the kernel about I/O utilization and scheduling behavior.

XXX - Come up with a better name than LP since this is working on
      big-core now.