WIP: PM QoS: Add PM_QOS_CPU_RESPONSE_FREQUENCY class.
The purpose of this PM QoS class is to give device drivers additional
control over the latency/energy efficiency trade-off made by the PM
subsystem (particularly the CPUFREQ governor).  It allows the device
driver to set a lower bound on the response latency of PM (defined as
the time it takes from wake-up to the CPU reaching a certain arbitrary
level of performance [e.g. the nominal frequency] in response to a
step-function load).  It reports to PM the minimum ramp-up latency
considered of use to the application, and explicitly requests PM to
filter out oscillations faster than the specified frequency.  It is
somewhat complementary to the current CPU_DMA_LATENCY PM QoS class
which can be understood as specifying an upper latency bound instead
of a lower bound.

Note that even though this provides a latency constraint it's
represented as its reciprocal in Hz units for computational efficiency
(since it would take a 64-bit division to compute the number of cycles
elapsed from a time increment in nanoseconds and a time bound, while a
frequency can simply be multiplied with the time increment).

This implements a MAX constraint so that the strictest (highest
response frequency) request is honored.  This means that PM won't
provide any guarantee that frequencies greater than the specified
bound will be filtered, since that might be incompatible with the
constraints specified by another more latency-sensitive application (A
more fine-grained result could be achieved with a scheduling-based
interface).  The default value needs to be equal to zero (best effort)
for it to behave as identity of the MAX operation.