WIP: cpufreq/intel_pstate: Implement energy-efficient controller target P-state range estimation.
The function introduced here calculates a P-state range derived from
the statistics computed in the previous patch which can be used to
drive the HWP P-state range directly or (if HWP is not available) as
basis for some additional kernel-side frequency selection mechanism
which will choose a single P-state from the range.  This is meant to
provide a reasonably energy-efficient frerquency range in cases where
the system has an IO bottleneck.