CHROMIUM: cpufreq: cpu-boost: Adapt the driver to the new world
When commit cb64fc8cd768 ("CHROMIUM: cpufreq: cpu-boost: Add cpu-boost
driver") was ported to Chrome OS 5.4 it was completely neutered.  The
driver was kept and the sysfs interface was still present, but it
didn't actually do anything.  This is because the way to enforce
cpufreq constraints changed and we didn't adapt.

Let's do the adaptation now.

NOTE: as mentioned in the bug associated with this change, we really
don't want to be carrying this long term.  We also have no test
coverage at all showing that this is beneficial.  These, of course,
need to be addressed.  However, while we're waiting for resources to
address them getting us back to the status quo of where we were on all
older kernels seems wise.

NOTE: from using a device without this patch I had a guess that typing
in gmail felt sluggish because this patch was missing.  Specifically,
if you fill up a window of text and then start typing at the top of
the email (so it has to word wrap / update for every character) you
could find that it was pretty sluggish.  After this patch it's not
sluggish.

NOTE ALSO: this patch needs to be combined with a change to the
overlay (the platform-cpusets.conf).  Specifically the input boost was
applied last in that script, but other errors (see chromium:1041117)
caused the script to exit before it got to the setting.

BUG=chromium:1032122
TEST=Typing in gmail ain't so sluggish no more.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ib96d739133fdb9ef0985e890d481b0d83be5cb23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2363953
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>