BACKPORT: FROMLIST: Bluetooth: Interleave with allowlist scan
This patch implements the interleaving between allowlist scan and
no-filter scan. It'll be used to save power when at least one monitor is
registered and at least one pending connection or one device to be
scanned for.

The durations of the allowlist scan and the no-filter scan are
controlled by MGMT command: Set Default System Configuration. The
default values are set randomly for now.

(am from https://patchwork.kernel.org/patch/11782041/)

BUG=b:163695127
TEST=add/remove adv_monitor through btmgmt and check the scanning type
is alternating with expected interval via btmon and debug logs.

conflicts:
- upstream has one blank line between hci_req_add_le_passive_scan and
  the following debug log. There is no such line in our repo.

Signed-off-by: Howard Chung <howardchung@google.com>
Signed-off-by: Alain Michaud <alainm@chromium.org>
Reviewed-by: Alain Michaud <alainm@chromium.org>
Reviewed-by: Manish Mandlik <mmandlik@chromium.org>
Change-Id: Ie5b3d5b9021faa014546fa35728ecef1eabfa9a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2416051