FROMGIT: drm/msm/dp: Sleep properly in dp_hpd_handler kthread
We shouldn't be waiting for an event here with a timeout of 100ms when
we're not in the 'timeout' arm of the if condition. Instead we should be
sleeping in the interruptible state (S) until something happens and we
need to wakeup. Right now this kthread is running almost all the time
because it sleeps for 100ms, wakes up, sees there's nothing to do, and
then starts the process all over again. Looking at top it shows up in
the D state (uninterruptible) because it uses wait_event_timeout(). FIx
this up.

Cc: Tanmay Shah <tanmay@codeaurora.org>
Cc: Kuogee Hsieh <khsieh@codeaurora.org>
Reported-by: Douglas Anderson <dianders@chromium.org>
Fixes: 8ede2ecc3e5e ("drm/msm/dp: Add DP compliance tests on Snapdragon Chipsets")
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Kuogee Hsieh <khsieh@codeaurora.org>
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 710a040a4dd7ff0e0b8df007c885e7f916a7b099
 https://gitlab.freedesktop.org/drm/msm.git msm-next)

BUG=b:153039340, b:168831365
TEST=build and boot lazor

Signed-off-by: Rob Clark <robdclark@chromium.org>
Change-Id: Ibdad92af90ef67ec5658245392ef4e4a06e23692
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2424535
Commit-Queue: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>