FROMGIT: usb: xhci: omit duplicate actions when suspending a runtime suspended host.
If the xhci-plat.c is the platform driver, after the runtime pm is
enabled, the xhci_suspend is called if nothing is connected on
the port. When the system goes to suspend, it will call xhci_suspend again
if USB wakeup is enabled.

Since the runtime suspend wakeup setting is not always the same as
system suspend wakeup setting, eg, at runtime suspend we always need
wakeup if the controller is in low power mode; but at system suspend,
we may not need wakeup. So, we move the judgement after changing
wakeup setting.

[commit message rewording -Mathias]

Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Link: https://lore.kernel.org/r/20200918131752.16488-8-mathias.nyman@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 18a367e8947d72dd91b6fc401e88a2952c6363f7
 git://git.kernel.org/pub/scm/linux/kernel/git/gregkh/usb.git usb-next)

BUG=b:120563698
TEST=build for trogdor

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I5402427d5351509da089fd6f78e3de85db1813bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2429631
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>